+++ crates/cli/src/lib.rs
@@ -9,7 +9,7 @@ use futures_core::TryStream;
 use futures_util::{TryStreamExt, pin_mut};
 use stac::{
     Assets, Collection, Item, Links, Migrate, SelfHref,
-    geoparquet::{Compression, DEFAULT_COMPRESSION},
+    geoparquet::{Compression, default_compression},
 };
 use stac_api::{GetItems, GetSearch, Search};
 use stac_io::{Format, StacStore};
@@ -593,7 +593,7 @@ impl Rustac {
             Format::Json(true)
         };
         if matches!(format, Format::Geoparquet(_)) {
-            Format::Geoparquet(self.parquet_compression.or(Some(DEFAULT_COMPRESSION)))
+            Format::Geoparquet(self.parquet_compression.or(Some(default_compression())))
         } else if let Format::Json(pretty) = format {
             Format::Json(self.compact_json.map(|c| !c).unwrap_or(pretty))
         } else {
@@ -784,7 +784,7 @@ mod tests {
     use assert_cmd::Command;
     use clap::Parser;
     use rstest::{fixture, rstest};
-    use stac::geoparquet::Compression;
+    use stac::geoparquet::{Compression, ZstdLevel};
     use stac_io::Format;
 
     #[fixture]
@@ -836,7 +836,7 @@ mod tests {
         let rustac = Rustac::parse_from(["Rustac", "translate"]);
         assert_eq!(
             rustac.input_format(Some("file.parquet")),
-            Format::Geoparquet(Some(Compression::SNAPPY))
+            Format::Geoparquet(Some(Compression::ZSTD(ZstdLevel::try_new(15).unwrap())))
         );
 
         let rustac = Rustac::parse_from(["rutsac", "--input-format", "json", "translate"]);
@@ -848,7 +848,7 @@ mod tests {
         let rustac = Rustac::parse_from(["rustac", "--input-format", "parquet", "translate"]);
         assert_eq!(
             rustac.input_format(None),
-            Format::Geoparquet(Some(Compression::SNAPPY))
+            Format::Geoparquet(Some(Compression::ZSTD(ZstdLevel::try_new(15).unwrap())))
         );
 
         let rustac = Rustac::parse_from([
@@ -876,7 +876,7 @@ mod tests {
         let rustac = Rustac::parse_from(["rustac", "translate"]);
         assert_eq!(
             rustac.output_format(Some("file.parquet")),
-            Format::Geoparquet(Some(Compression::SNAPPY))
+            Format::Geoparquet(Some(Compression::ZSTD(ZstdLevel::try_new(15).unwrap())))
         );
 
         let rustac = Rustac::parse_from(["rustac", "--output-format", "json", "translate"]);
@@ -888,7 +888,7 @@ mod tests {
         let rustac = Rustac::parse_from(["rustac", "--output-format", "parquet", "translate"]);
         assert_eq!(
             rustac.output_format(None),
-            Format::Geoparquet(Some(Compression::SNAPPY))
+            Format::Geoparquet(Some(Compression::ZSTD(ZstdLevel::try_new(15).unwrap())))
         );
 
         let rustac = Rustac::parse_from([