+++ crates/core/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+- Switch to zstd(15) for default geoparquet compression ([#834](https://github.com/stac-utils/rustac/pull/834))
+
 ## [0.13.2] - 2025-10-22
 
 ### Fixed